requirements.txt
bottle==0.12.9
facebook-sdk==1.0.0
requests==2.9.1
wsgiref==0.1.2
